feat: separate ledger testcase to testnet
